# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoreBanking API
  version: 1.0.0-phase1
  description: |
    Staff and integration API for CoreBanking. Every request is tenant-scoped: the tenant comes from the
    verified access token (`tenant` claim), never from the URL or body.

    Conventions
    - Money is a decimal **string** (e.g. "100000.00"). Never a JSON number.
    - Dates are ISO `YYYY-MM-DD` business dates; timestamps are RFC 3339 UTC.
    - Changes to master data and money go through maker-checker: the call returns **202** with an
      approval request; the change is applied when a different user approves it.
    - Errors use RFC 9457 `application/problem+json`.
    - `Idempotency-Key` header is honoured on all POSTs that create things.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearer: []
...
paths:
...
  /api/v1/approvals/{id}/approve:
    post:
      tags: [Approvals]
      operationId: approve
      summary: Approve
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Decision' }
      responses:
        '200':
          description: Approved and applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/sessions:
    get:
      tags: [Session]
      operationId: listSessions
      summary: The caller's login sessions, or another user's with session:admin
      parameters:
        - { name: user, in: query, description: User name; needs `session:admin`, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LoginSession' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sessions/{id}:
    delete:
      tags: [Session]
      operationId: terminateSession
      summary: End a login session
      description: >-
        Without `user` the session must be one of the caller's own; with `user` (`session:admin`) one of that user's.
        Every termination is audited. An access token already issued stays valid until it expires.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: user, in: query, schema: { type: string } }
      responses:
        '204': { description: Ended }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: Job catalogue with schedules, next fire time and last run
      description: Permission `job:view`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Job' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/jobs/runs:
    get:
      tags: [Jobs]
      operationId: listJobRuns
      summary: Job runs, newest first
      parameters:
        - { name: job, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/JobRun' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/jobs/{code}:
    put:
      tags: [Jobs]
      operationId: proposeJobSchedule
      summary: Propose a job's schedule, on/off switch and parameters (maker-checker)
      description: >-
        Permission `job:schedule`. The schedule is a six-field cron (second minute hour day-of-month month day-of-week)
        in India Standard Time. For a report job the maker needs the report's permission; the scheduled runs use the
        maker's branch scope and belong to the maker.
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobScheduleInput' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/jobs/{code}/run:
    post:
      tags: [Jobs]
      operationId: runJob
      summary: Run a job now
      description: Permission `job:run`. Answers when the job has finished. 409 when the job is already running.
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRun' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/support-access:
    get:
      tags: [Platform]
      operationId: listSupportAccess
      summary: Support access requests and grants of this tenant
      description: Permission `support-access:approve`.
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [REQUESTED, APPROVED, REJECTED, REVOKED, EXPIRED, LAPSED] } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SupportAccess' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/support-access/{id}/approve:
    post:
      tags: [Platform]
      operationId: approveSupportAccess
      summary: Approve a support access request; it expires after the requested duration
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Note' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportAccess' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/support-access/{id}/reject:
    post:
      tags: [Platform]
      operationId: rejectSupportAccess
      summary: Reject a support access request (a note is required)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Note' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportAccess' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/support-access/{id}/revoke:
    post:
      tags: [Platform]
      operationId: revokeSupportAccess
      summary: Revoke a support access grant at once
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Note' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportAccess' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Accepted:
      description: Change recorded as an approval request; applied when a checker approves it
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Approval' }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Money:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,4})?$'
      examples: ['100000.00']
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Field-level problems (validation), when there are several
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, WITHDRAWN]
    Approval:
      type: object
      required: [id, entityType, action, status, maker, madeAt]
      properties:
        id: { type: string, format: uuid }
        entityType: { type: string, examples: [CUSTOMER, BRANCH, BRANCH_SET, STAFF, GL_HEAD, VOUCHER, TAX_RATE, HOLIDAY, TERRITORY, ENUMERATION, SYSTEM_PROPERTY, EOD_SCHEDULE, LOAN_PRODUCT, LOAN_DISBURSEMENT, LOAN_WAIVER, LOAN_REVERSAL, LOAN_AMENDMENT, LOAN_RESTRUCTURE] }
        entityId: { type: [string, 'null'] }
        action: { type: string, examples: [CREATE, UPDATE, REVERSE] }
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        maker: { type: string }
        madeAt: { type: string, format: date-time }
        checker: { type: [string, 'null'] }
        checkedAt: { type: [string, 'null'], format: date-time }
        note: { type: [string, 'null'] }
        amount: { oneOf: [ { $ref: '#/components/schemas/Money' }, { type: 'null' } ] }
        ageHours: { type: number, description: SLA ageing }
        checkersRequired: { type: integer, description: 'Distinct checkers needed (0-3), from the approval rules' }
        approvalsSoFar: { type: integer, description: Approvals already recorded while PENDING }
        appliedRef: { type: [string, 'null'], description: 'What the approval produced (customer id, voucher number …)' }
        current: { type: [object, 'null'], description: State before the change (null for CREATE) }
        proposed: { type: object, description: Proposed state }
    Decision:
      type: object
      properties:
        note: { type: string, maxLength: 500 }
...
    Note:
      type: object
      properties:
        note: { type: string }
...
    LoginSession:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        ipAddress: { type: [string, 'null'] }
        startedAt: { type: [string, 'null'], format: date-time }
        lastAccessAt: { type: [string, 'null'], format: date-time }
        clients: { type: array, items: { type: string } }
        current: { type: boolean, description: The session this request's token belongs to }
    Job:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        kind: { type: string, enum: [REPORT, DASHBOARD_REFRESH, KYC_EXPIRY, CONSENT_EXPIRY, EXPORT_CLEANUP, USAGE_SNAPSHOT, DEFERRED_RECEIPTS] }
        schedule: { type: [string, 'null'], description: Six-field cron in IST }
        enabled: { type: boolean }
        parameters: { type: object, additionalProperties: true }
        builtIn: { type: boolean }
        updatedBy: { type: string }
        updatedAt: { type: string, format: date-time }
        runnable: { type: boolean }
        nextRunAt: { type: [string, 'null'], format: date-time }
        lastRun:
          oneOf:
            - $ref: '#/components/schemas/JobRun'
            - type: 'null'
    JobRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        jobCode: { type: string }
        origin: { type: string, enum: [SCHEDULE, MANUAL] }
        scheduledFor: { type: [string, 'null'], format: date-time }
        requestedBy: { type: string }
        status: { type: string, enum: [RUNNING, COMPLETED, FAILED] }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        processed: { type: integer }
        failed: { type: integer }
        error: { type: [string, 'null'] }
        artifact:
          type: [object, 'null']
          additionalProperties: true
          description: What the run produced. A scheduled report carries `reportRunId` and `delivery` (PENDING_PROVIDER until e-mail is built).
    JobScheduleInput:
      type: object
      properties:
        schedule: { type: [string, 'null'], examples: ['0 0 6 1 * *'] }
        enabled: { type: boolean }
        parameters:
          type: object
          additionalProperties: true
          description: >-
            Report jobs take `period` (BUSINESS_DATE, PREVIOUS_DAY, MONTH_TO_DATE, PREVIOUS_MONTH), `parameters` (the
            report's own, without dates) and `emailTo`. Other jobs take none.
...
    SupportAccess:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant: { type: string, description: Present in control-plane responses }
        engineer: { type: string }
        reason: { type: string }
        ticket: { type: string }
        scope: { type: string }
        durationMinutes: { type: integer }
        requestedAt: { type: string, format: date-time }
        status: { type: string, enum: [REQUESTED, APPROVED, REJECTED, REVOKED, EXPIRED, LAPSED, UNKNOWN] }
        decidedBy: { type: [string, 'null'] }
        decidedAt: { type: [string, 'null'], format: date-time }
        decisionNote: { type: [string, 'null'] }
        expiresAt: { type: [string, 'null'], format: date-time }
        revokedBy: { type: [string, 'null'] }
        revokedAt: { type: [string, 'null'], format: date-time }
        revokeReason: { type: [string, 'null'] }
...